Move fetchData to actions and update types
