+++ src/actions/index.tsx
@@ -1,5 +1,7 @@
 import checkoutAction from './checkout'
 import billingAction from './billing'
+import fetchDataAction from './fetchData'
 
 export const checkout = checkoutAction
 export const billing = billingAction
+export const fetchData = fetchDataAction
